push

+++ backend/controllers/supplierControllers.js
@@ -1,4 +1,5 @@
 const Supplier = require('../models/supplierModel')
+const Stock = require('../models/stockModel')
 const mongoose = require('mongoose')
 const removeSpecialChar = require('../helpers/helper')
 
@@ -12,8 +13,7 @@ const getAllSuppliers = async(req, res) => {
         if (name)
             filter.name = { $regex: `.*${removeSpecialChar(name)}.*`, $options: 'i' }
 
-        const suppliers = await Supplier.find(filter)
-            .sort({ createdAt: -1 })
+        const suppliers = await Supplier.find({ ...filter, isDelete: false }).sort({ createdAt: -1 })
   
         res.status(200).json(suppliers)
     } 
@@ -27,8 +27,7 @@ const getSupplier = async (req, res) => {
     const { id } = req.params
 
     try {
-        const supplier = await Supplier.findById(id)
-            .sort({ createdAt: -1 })
+        const supplier = await Supplier.findOne({_id: id, isDelete: false})
   
         res.status(200).json(supplier)
     } 
@@ -61,11 +60,19 @@ const deleteSupplier = async (req, res) => {
     if (!mongoose.Types.ObjectId.isValid(id))
         return res.status(400).json({error: 'No such supplier'})
     try {
-        const supplier = await Supplier.findByIdAndDelete(id)
+        let supplier = await Supplier.findById(id)
 
-        if(!supplier)
-            return res.status(400).json({error: 'No such supplier'})
-    
+        if (!supplier || supplier.isDelete)
+            return res.status(404).json({error: 'Supplier not found or already deleted'})
+
+        const stock = await Stock.findOne({'items.supplierID': id})
+
+        if (stock)
+            supplier = await Supplier.findByIdAndUpdate(id, { isDelete: true }, { new: true, runValidators: true })
+
+        else
+            supplier = await Supplier.findByIdAndDelete(id)
+
         res.status(200).json(supplier)
     }
     catch (error) {
@@ -79,15 +86,17 @@ const updateSupplier = async (req, res) => {
 
     if (!mongoose.Types.ObjectId.isValid(id))
         res.status(400).json({error: 'No such supplier'})
-    
-    try {
-        const supplier = await Supplier.findByIdAndUpdate(id, req.body, { new: true, runValidators: true })
 
-        if(!supplier)
-            res.status(400).json({ error: 'No such supplier' })
-    
+    try {
+        let supplier = await Supplier.findById(id)
+
+        if (!supplier || supplier.isDelete)
+            return res.status(404).json({error: 'Supplier not found or already deleted'})
+
+        supplier = await Supplier.findByIdAndUpdate(id, req.body, { new: true, runValidators: true })
+
         res.status(200).json(supplier)
-    }
+    } 
     catch (error) {
         res.status(500).json({error: error.message})
     }
